config/gen_from_riscv_config/requirements.txt
docutils
pyyaml
mdutils
restructuredtext-lint
rstcloth
regex